requirements.txt
absl-py==1.0.0
astunparse==1.6.3
beautifulsoup4==4.8.0
bson==0.5.8
cached-property==1.5.2
cachetools==4.2.4
cassandra-driver==3.19.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cql==1.4.0
cycler==0.10.0
dataclasses==0.6
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.4
django-cassandra-engine==1.5.5
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.4.0
django-import-export==2.5.0
django-mathfilters==1.0.0
django-mongodb-engine==0.6.0
django-sslserver==0.22
djangorestframework==3.11.0
djangotoolbox==1.8.0
djongo==1.3.1
et-xmlfile==1.0.1
Flask==1.1.1
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googlemaps==4.4.2
GoogleNews==1.4.0
grpcio==1.43.0
h5py==3.6.0
idna==2.8
importlib-metadata==4.10.0
instamojo-wrapper==1.2.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
joblib==0.14.0
Js2Py==0.70
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
libclang==12.0.0
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.1.2
mysql-connector==2.2.9
mysqlclient==1.4.4
netifaces==0.10.9
nltk==3.5
numpy==1.17.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
opt-einsum==3.3.0
pandas==1.2.2
Pillow==6.1.0
pipenv==2018.11.26
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pyjsparser==2.7.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.5
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.2
PyYAML==5.4.1
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
reverse-geocoder==1.5.1
rsa==4.8
scikit-learn==0.21.3
scipy==1.3.3
selenium==3.141.0
sentry-sdk==1.3.1
six==1.13.0
sklearn==0.0
soupsieve==1.9.2
sqlparse==0.2.4
tablib==3.0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
textblob==0.15.3
thrift==0.13.0
tqdm==4.48.2
tweepy==3.9.0
typing-extensions==4.0.1
tzlocal==2.0.0
urllib3==1.25.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
Werkzeug==0.15.5
wikipedia==1.4.0
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
